Removal of unused files due to refactoring
